\name{color_bar}
\alias{color_bar}

\title{
 generate color bar
}

\description{
 generate color bar
}

\usage{
  color_bar(cols=cols, val=val, coord=coord, lab.s=lab.s, cex=0.4, horizontal=T);
}

\arguments{
   \item{cols}{cols: colors}
   \item{val}{val: values}
   \item{coord}{coord: x1, y1, x2, y2}
   \item{lab.s}{lab.s: adjust label}
   \item{cex}{cex: }
   \item{horizontal}{horizontal:}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

