\name{doGSEA}
\alias{doGSEA}

\title{
 do GSEA
}

\description{
 do GSEA
}

\usage{
  doGSEA(db=db, gene=gene, filter.num=0, fdr=FALSE);
}

\arguments{
   \item{db}{database name: one of them: GO, KEGG, ...}
   \item{gene}{vector: gene symbols}
   \item{filter.num}{min gene number}
   \item{fdr}{do fdr p value}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

