\name{do_neighborhood}
\alias{do_neighborhood}

\title{
 make networks by gene neighborhoods
}

\description{
 make networks by gene neighborhoods
}

\usage{
  do_neighborhood(g=g, genes=genes, order=1)
}

\arguments{
   \item{g}{graph}
   \item{genes}{vector: gene symbols}
   \item{order}{Integer giving the order of the neighborhood}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

