\name{ES}
\alias{ES}

\title{
 calculate enrichment score and p value
}

\description{
 calculate enrichment score and p value
}

\usage{
  ES(db=db, gene=gene, pdffile=pdffile, fdr=T);
}

\arguments{
   \item{db}{db: a numeric vector of your scores (prevalence of variants) of all genes in your dataset}
   \item{gene}{a vector of genes}
   \item{pdffile}{output pdf file name}
   \item{fdr}{logical: TURE or FALSE, do fdr pvalue}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

