\name{hs2mm}
\alias{hs2mm}

\title{
 convert human gene symbols to mouse
}

\description{
 convert human gene symbols to mouse
}

\usage{
  hs2mm(gene=gene);
}

\arguments{
   \item{gene}{vector: gene symbols}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

