\name{KEGGplot}
\alias{KEGGplot}

\title{
 KEGGplot
}

\description{
 KEGGplot
}

\usage{
  KEGGplot(pid=pid, Taxonomy="hs", plot=TRUE);
}

\arguments{
   \item{pid}{pid: KEGG}
   \item{Taxonomy}{Taxonomy: hs or mm}
   \item{plot}{plot out}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

