\name{mm2hs}
\alias{mm2hs}

\title{
 convert mouse gene symbols to human
}

\description{
 convert mouse gene symbols to human
}

\usage{
  mm2hs(gene=gene);
}

\arguments{
   \item{gene}{vector: gene symbols}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

