%% main
\name{OmicPath-package}
\alias{OmicPath}
\title{
 OmicPath package
}
\description{
 OmicPath tools.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}
\references{
%%  ~~ possibly secondary sources and usages ~~
 OmicPath: an R package for analysis and visulazation of pathway and bio-network
}
%%\keyword{key}
