\name{pathway_creator}
\alias{pathway_creator}

\title{
 pathway_creator
}

\description{
 pathway_creator
}

\usage{
  pathway_creator(gene.table=gene.table, creator=creator, 
                  version=version, mydate=NULL, dbname="GSEA.db");
}

\arguments{
   \item{gene.table}{data frame: column 1: pathway name, column 2: gene list and column 3: pathway name}
   \item{creator}{string: name}
   \item{version}{string: db version}
   \item{mydate}{time: time}
   \item{dbname}{pathway database name}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

