\name{pathway_delete}
\alias{pathway_delete}

\title{
 pathway_delete
}

\description{
 pathway_delete
}

\usage{
  pathway_delete(db=db, pathway.ID=pathway.ID,
                           creator=creator, 
                           version=version, mydate=NULL);
}

\arguments{
   \item{db}{database name: one of them: GO, KEGG, ...}
   \item{pathway.ID}{string: pathway symbol}
   \item{creator}{string: name}
   \item{version}{string: db version}
   \item{mydate}{time: time}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

