\name{pathway_search}
\alias{pathway_search}

\title{
 pathway_search
}

\description{
 pathway_search
}

\usage{
  pathway_search(db=db, gene=gene, pathway.ID=NULL);
}

\arguments{
   \item{db}{database name: one of them: GO, KEGG, ...}
   \item{gene}{vector: gene symbols}
   \item{pathway.ID}{string: pathway symbol}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

