\name{pathway_update}
\alias{pathway_update}

\title{
 pathway_update
}

\description{
 pathway_update
}

\usage{
  pathway_update(db=db, gene.list=gene.list, pathway.ID=pathway.ID,
                           pathway.name=pathway.name, creator=creator, 
                           version=version, mydate=NULL);
}

\arguments{
   \item{db}{database name: one of them: GO, KEGG, ...}
   \item{gene.list}{vector: gene list}
   \item{pathway.ID}{string: pathway symbol}
   \item{pathway.name}{string: pathway name}
   \item{creator}{string: name}
   \item{version}{string: db version}
   \item{mydate}{time: time}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

