\name{plotKEGG}
\alias{plotKEGG}

\title{
 plotKEGG
}

\description{
 plotKEGG
}

\usage{
  plotKEGG(pid=pid, pname=pname, Taxonomy="hs", gene=gene, glab=F, fc.value=fc.value, 
                     p.value=p.value, pdffile=pdffile);
}

\arguments{
   \item{pid}{pid: KEGG}
   \item{pname}{pname: pathway name}
   \item{Taxonomy}{Taxonomy: hs or mm}
   \item{gene}{gene: genes}
   \item{glab}{glab: lablel gene}
   \item{fc.value}{fc.value: fc value}
   \item{p.value}{p.value: p value}
   \item{pdffile}{pdffile: pdffile}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

