\name{plotNet}
\alias{plotNet}

\title{
 plotNet
}

\description{
 plotNet
}

\usage{
  plotNet(db=db, dat=dat, Taxonomy="mm", cluster=TRUE, pdffile=pdffile);
}

\arguments{
   \item{db}{db: data base}
   \item{dat}{dat: column 1: gene}
   \item{Taxonomy}{Taxonomy: hs or mm}
   \item{cluster}{cluster: cluster analysis}
   \item{pdffile}{pdffile: pdffile}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

