\name{similarity}
\alias{similarity}

\title{
 do similarity
}

\description{
 do similarity
}

\usage{
  similarity(s1, s2);
}

\arguments{
   \item{s1}{vector: set 1}
   \item{s2}{vector: set 2}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

