\name{steiner_net}
\alias{steiner_net}

\title{
 do steiner_net
}

\description{
 do steiner_net
}

\usage{
  steiner_net(g=g, steiner.node=steiner.node, plot.net=F);
}

\arguments{
   \item{g}{igraph object}
   \item{steiner.node}{vector: node (steiner points)}
   \item{plot.net}{logical: print if TRUE}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

