\name{WGCNA_6_1}
\alias{WGCNA_6_1}

\title{
 Step 1: checking data for WGCNA 
}

\description{
 Step 1: checking data for WGCNA 
}

\usage{
  WGCNA_6_1(expr=expr, pheno=pheno, cutoff=8000, cutHeight=200, project.name="WGCNA_net");
}

\arguments{
   \item{expr}{expr: gene expression data}
   \item{pheno}{pheno: phenotype data}
   \item{cutoff}{cutoff: selecting top variance gene number}
   \item{cutHeight}{cutHeight: outlier threshold}
   \item{project.name}{project.name: project name}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

