\name{WGCNA_6_2}
\alias{WGCNA_6_2}

\title{
 Step 2: estmating power value for WGCNA 
}

\description{
 Step 2: estmating power value for WGCNA 
}

\usage{
  WGCNA_6_2(expr=expr, project.name="WGCNA_net");
}

\arguments{
   \item{expr}{expr: gene expression data}
   \item{project.name}{project.name: project name}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

