\name{WGCNA_6_3}
\alias{WGCNA_6_3}

\title{
 Step 3: determing module size for WGCNA 
}

\description{
 Step 3: determing module size for WGCNA 
}

\usage{
  WGCNA_6_3(expr=expr, softPower=8, minM=minM, cutH=cutH, stype="unsigned", project.name="WGCNA_net");
}

\arguments{
   \item{expr}{expr: gene expression data}
   \item{softPower}{softPower: power value from step 2}
   \item{minM}{minM: minimum node number per module}
   \item{cutH}{cutH: threshold for module selecting}
   \item{stype}{stype: unsigned, signed, signed hybrid, distance}
   \item{project.name}{project.name: project name}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

