\name{WGCNA_6_5}
\alias{WGCNA_6_5}

\title{
 Step 5: generating network for WGCNA 
}

\description{
 Step 5: generating network for WGCNA 
}

\usage{
  WGCNA_6_5(mdir="module", project.name="WGCNA_net", cutoff=20000);
}

\arguments{
   \item{mdir}{mdir: output directory for module}
   \item{project.name}{project.name: project name}
   \item{cutoff}{cutoff: gene number >= the cutoff, the module will be ignored.}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

