\name{WGCNA_6_6}
\alias{WGCNA_6_6}

\title{
 Step 6: GSEA for module for WGCNA 
}

\description{
 Step 6: GSEA for module for WGCNA 
}

\usage{
  WGCNA_6_6(gdir="GSEA", mdir="module", db=db);
}

\arguments{
   \item{gdir}{gdir: output directory for GSEA}
   \item{mdir}{mdir: output directory for module}
   \item{db}{db: gene sets for GSEA}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

