\name{WGCNA_dissTOM}
\alias{WGCNA_dissTOM}

\title{
 get dissTOM matrix
}

\description{
 get dissTOM matrix
}

\usage{
  WGCNA_dissTOM(tom=tom, modules=modules, indir=indir, outdir=outdir);
}

\arguments{
   \item{tom}{dissTOM matrix}
   \item{modules}{module file vector}
   \item{indir}{input directory}
   \item{outdir}{output directory}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

