\name{WGCNA_dissTOM2network}
\alias{WGCNA_dissTOM2network}

\title{
 get network from dissTOM matrix
}

\description{
 get network from dissTOM matrix
}

\usage{
  WGCNA_dissTOM2network(m.files=m.files, indir=indir, outdir=outdir, project.name="ALL", softPower=6,
     cutoff=cutoff, label=FALSE, label.cex=0.4);
}

\arguments{
   \item{m.files}{files of tom matrix}
   \item{indir}{input directory}
   \item{outdir}{output directory}
   \item{project.name}{project name}
   \item{softPower}{softPower}
   \item{cutoff}{cutoff}
   \item{label}{logical, gene lable}
   \item{label.cex}{gene lable size}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

